+++ tests/nlu/featurizers/test_count_vectors_featurizer.py
@@ -1,255 +1,9 @@
-# -*- coding: utf-8 -
 import numpy as np
 import pytest
 
-from rasa.nlu import training_data
 from rasa.nlu.tokenizers.tokenizer import Token
-from rasa.nlu.tokenizers.mitie_tokenizer import MitieTokenizer
-from rasa.nlu.tokenizers.spacy_tokenizer import SpacyTokenizer
 from rasa.nlu.training_data import Message
 from rasa.nlu.training_data import TrainingData
-from rasa.nlu.config import RasaNLUModelConfig
-
-
-@pytest.mark.parametrize(
-    "sentence, expected",
-    [
-        (
-            "hey how are you today",
-            [-0.19649599, 0.32493639, -0.37408298, -0.10622784, 0.062756],
-        )
-    ],
-)
-def test_spacy_featurizer(sentence, expected, spacy_nlp):
-    from nlu.featurizers.dense_featurizer import spacy_featurizer
-
-    doc = spacy_nlp(sentence)
-    vecs = spacy_featurizer.features_for_doc(doc)
-    assert np.allclose(doc.vector[:5], expected, atol=1e-5)
-    assert np.allclose(vecs, doc.vector, atol=1e-5)
-
-
-def test_spacy_training_sample_alignment(spacy_nlp_component):
-    from spacy.tokens import Doc
-
-    m1 = Message.build(text="I have a feeling", intent="feeling")
-    m2 = Message.build(text="", intent="feeling")
-    m3 = Message.build(text="I am the last message", intent="feeling")
-    td = TrainingData(training_examples=[m1, m2, m3])
-
-    attribute_docs = spacy_nlp_component.docs_for_training_data(td)
-
-    assert isinstance(attribute_docs["text"][0], Doc)
-    assert isinstance(attribute_docs["text"][1], Doc)
-    assert isinstance(attribute_docs["text"][2], Doc)
-
-    assert [t.text for t in attribute_docs["text"][0]] == ["i", "have", "a", "feeling"]
-    assert [t.text for t in attribute_docs["text"][1]] == []
-    assert [t.text for t in attribute_docs["text"][2]] == [
-        "i",
-        "am",
-        "the",
-        "last",
-        "message",
-    ]
-
-
-def test_spacy_intent_featurizer(spacy_nlp_component):
-    from nlu.featurizers.dense_featurizer.spacy_featurizer import SpacyFeaturizer
-
-    td = training_data.load_data("data/examples/rasa/demo-rasa.json")
-    spacy_nlp_component.train(td, config=None)
-    spacy_featurizer = SpacyFeaturizer()
-    spacy_featurizer.train(td, config=None)
-
-    intent_features_exist = np.array(
-        [
-            True if example.get("intent_features") is not None else False
-            for example in td.intent_examples
-        ]
-    )
-
-    # no intent features should have been set
-    assert not any(intent_features_exist)
-
-
-@pytest.mark.parametrize(
-    "sentence, expected",
-    [("hey how are you today", [-0.28451, 0.31007, -0.57039, -0.073056, -0.17322])],
-)
-def test_spacy_ner_featurizer(sentence, expected, spacy_nlp):
-    from nlu.featurizers.dense_featurizer.spacy_featurizer import SpacyFeaturizer
-
-    doc = spacy_nlp(sentence)
-    token_vectors = [t.vector for t in doc]
-    spacy_config = {"ner_feature_vectors": True}
-    ftr = SpacyFeaturizer.create(spacy_config, RasaNLUModelConfig())
-    greet = {"intent": "greet", "text_features": [0.5]}
-    message = Message(sentence, greet)
-    message.set("spacy_doc", doc)
-    ftr._set_spacy_features(message)
-    ftr._set_spacy_ner_features(message)
-    vecs = message.get("ner_features")[0][:5]
-    assert np.allclose(token_vectors[0][:5], vecs, atol=1e-4)
-    assert np.allclose(vecs, expected, atol=1e-4)
-
-
-def test_spacy_ner_featurizer_config(spacy_nlp):
-    from nlu.featurizers.dense_featurizer.spacy_featurizer import SpacyFeaturizer
-
-    sentence = "hi there friend"
-    doc = spacy_nlp(sentence)
-    spacy_config = {"ner_feature_vectors": False}
-    ftr = SpacyFeaturizer.create(spacy_config, RasaNLUModelConfig())
-    greet = {"intent": "greet", "text_features": [0.5]}
-    message = Message(sentence, greet)
-    message.set("spacy_doc", doc)
-    ftr._set_spacy_features(message)
-    ftr._set_spacy_ner_features(message)
-    vecs = np.array(message.get("ner_features"))
-    assert vecs.shape[0] == len(doc)
-    assert vecs.shape[1] == 0
-
-
-def test_mitie_featurizer(mitie_feature_extractor, default_config):
-    from nlu.featurizers.dense_featurizer.mitie_featurizer import MitieFeaturizer
-
-    mitie_component_config = {"name": "MitieFeaturizer"}
-    ftr = MitieFeaturizer.create(mitie_component_config, RasaNLUModelConfig())
-    sentence = "Hey how are you today"
-    tokens = MitieTokenizer().tokenize(sentence)
-    vecs = ftr.features_for_tokens(tokens, mitie_feature_extractor)
-    expected = np.array([0.0, -4.4551446, 0.26073121, -1.46632245, -1.84205751])
-    assert np.allclose(vecs[:5], expected, atol=1e-5)
-
-
-def test_ngram_featurizer(spacy_nlp):
-    from nlu.featurizers.sparse_featurizer.ngram_featurizer import NGramFeaturizer
-
-    ftr = NGramFeaturizer({"max_number_of_ngrams": 10})
-
-    # ensures that during random sampling of the ngram CV we don't end up
-    # with a one-class-split
-    repetition_factor = 5
-
-    greet = {"intent": "greet", "text_features": [0.5]}
-    goodbye = {"intent": "goodbye", "text_features": [0.5]}
-    labeled_sentences = [
-        Message("heyheyheyhey", greet),
-        Message("howdyheyhowdy", greet),
-        Message("heyhey howdyheyhowdy", greet),
-        Message("howdyheyhowdy heyhey", greet),
-        Message("astalavistasista", goodbye),
-        Message("astalavistasista sistala", goodbye),
-        Message("sistala astalavistasista", goodbye),
-    ] * repetition_factor
-
-    for m in labeled_sentences:
-        m.set("spacy_doc", spacy_nlp(m.text))
-
-    ftr.min_intent_examples_for_ngram_classification = 2
-    ftr.train_on_sentences(labeled_sentences)
-    assert len(ftr.all_ngrams) > 0
-    assert ftr.best_num_ngrams > 0
-
-
-@pytest.mark.parametrize(
-    "sentence, expected, labeled_tokens",
-    [
-        ("hey how are you today", [0.0, 1.0, 0.0], [0]),
-        ("hey 456 how are you", [1.0, 1.0, 0.0], [1, 0]),
-        ("blah balh random eh", [0.0, 0.0, 0.0], []),
-        ("a 1 digit number", [1.0, 0.0, 1.0], [1, 1]),
-    ],
-)
-def test_regex_featurizer(sentence, expected, labeled_tokens, spacy_nlp):
-    from nlu.featurizers.sparse_featurizer.regex_featurizer import RegexFeaturizer
-
-    patterns = [
-        {"pattern": "[0-9]+", "name": "number", "usage": "intent"},
-        {"pattern": "\\bhey*", "name": "hello", "usage": "intent"},
-        {"pattern": "[0-1]+", "name": "binary", "usage": "intent"},
-    ]
-    ftr = RegexFeaturizer(known_patterns=patterns)
-
-    # adds tokens to the message
-    tokenizer = SpacyTokenizer()
-    message = Message(sentence)
-    message.set("spacy_doc", spacy_nlp(sentence))
-    tokenizer.process(message)
-
-    result = ftr.features_for_patterns(message)
-    assert np.allclose(result, expected, atol=1e-10)
-
-    # the tokenizer should have added tokens
-    assert len(message.get("tokens", [])) > 0
-    # the number of regex matches on each token should match
-    for i, token in enumerate(message.get("tokens")):
-        token_matches = token.get("pattern").values()
-        num_matches = sum(token_matches)
-        assert num_matches == labeled_tokens.count(i)
-
-
-@pytest.mark.parametrize(
-    "sentence, expected, labeled_tokens",
-    [
-        ("lemonade and mapo tofu", [1, 1], [0.0, 2.0, 3.0]),
-        ("a cup of tea", [1, 0], [3.0]),
-        ("Is burrito my favorite food?", [0, 1], [1.0]),
-        ("I want club?mate", [1, 0], [2.0, 3.0]),
-    ],
-)
-def test_lookup_tables(sentence, expected, labeled_tokens, spacy_nlp):
-    from nlu.featurizers.sparse_featurizer.regex_featurizer import RegexFeaturizer
-
-    lookups = [
-        {
-            "name": "drinks",
-            "elements": ["mojito", "lemonade", "sweet berry wine", "tea", "club?mate"],
-        },
-        {"name": "plates", "elements": "data/test/lookup_tables/plates.txt"},
-    ]
-    ftr = RegexFeaturizer(lookup_tables=lookups)
-
-    # adds tokens to the message
-    tokenizer = SpacyTokenizer()
-    message = Message(sentence)
-    message.set("spacy_doc", spacy_nlp(sentence))
-    tokenizer.process(message)
-
-    result = ftr.features_for_patterns(message)
-    assert np.allclose(result, expected, atol=1e-10)
-
-    # the tokenizer should have added tokens
-    assert len(message.get("tokens", [])) > 0
-    # the number of regex matches on each token should match
-    for i, token in enumerate(message.get("tokens")):
-        token_matches = token.get("pattern").values()
-        num_matches = sum(token_matches)
-        assert num_matches == labeled_tokens.count(i)
-
-
-def test_spacy_featurizer_casing(spacy_nlp):
-    from nlu.featurizers.dense_featurizer import spacy_featurizer
-
-    # if this starts failing for the default model, we should think about
-    # removing the lower casing the spacy nlp component does when it
-    # retrieves vectors. For compressed spacy models (e.g. models
-    # ending in _sm) this test will most likely fail.
-
-    td = training_data.load_data("data/examples/rasa/demo-rasa.json")
-    for e in td.intent_examples:
-        doc = spacy_nlp(e.text)
-        doc_capitalized = spacy_nlp(e.text.capitalize())
-
-        vecs = spacy_featurizer.features_for_doc(doc)
-        vecs_capitalized = spacy_featurizer.features_for_doc(doc_capitalized)
-
-        assert np.allclose(
-            vecs, vecs_capitalized, atol=1e-5
-        ), "Vectors are unequal for texts '{}' and '{}'".format(
-            e.text, e.text.capitalize()
-        )
 
 
 @pytest.mark.parametrize(
