+++ tests/nlu/featurizers/test_regex_featurizer.py
@@ -1,0 +1,86 @@
+import numpy as np
+import pytest
+
+from rasa.nlu import training_data
+from rasa.nlu.tokenizers.tokenizer import Token
+from rasa.nlu.tokenizers.mitie_tokenizer import MitieTokenizer
+from rasa.nlu.tokenizers.spacy_tokenizer import SpacyTokenizer
+from rasa.nlu.training_data import Message
+from rasa.nlu.training_data import TrainingData
+from rasa.nlu.config import RasaNLUModelConfig
+
+
+@pytest.mark.parametrize(
+    "sentence, expected, labeled_tokens",
+    [
+        ("hey how are you today", [0.0, 1.0, 0.0], [0]),
+        ("hey 456 how are you", [1.0, 1.0, 0.0], [1, 0]),
+        ("blah balh random eh", [0.0, 0.0, 0.0], []),
+        ("a 1 digit number", [1.0, 0.0, 1.0], [1, 1]),
+    ],
+)
+def test_regex_featurizer(sentence, expected, labeled_tokens, spacy_nlp):
+    from nlu.featurizers.sparse_featurizer.regex_featurizer import RegexFeaturizer
+
+    patterns = [
+        {"pattern": "[0-9]+", "name": "number", "usage": "intent"},
+        {"pattern": "\\bhey*", "name": "hello", "usage": "intent"},
+        {"pattern": "[0-1]+", "name": "binary", "usage": "intent"},
+    ]
+    ftr = RegexFeaturizer(known_patterns=patterns)
+
+    # adds tokens to the message
+    tokenizer = SpacyTokenizer()
+    message = Message(sentence)
+    message.set("spacy_doc", spacy_nlp(sentence))
+    tokenizer.process(message)
+
+    result = ftr.features_for_patterns(message)
+    assert np.allclose(result, expected, atol=1e-10)
+
+    # the tokenizer should have added tokens
+    assert len(message.get("tokens", [])) > 0
+    # the number of regex matches on each token should match
+    for i, token in enumerate(message.get("tokens")):
+        token_matches = token.get("pattern").values()
+        num_matches = sum(token_matches)
+        assert num_matches == labeled_tokens.count(i)
+
+
+@pytest.mark.parametrize(
+    "sentence, expected, labeled_tokens",
+    [
+        ("lemonade and mapo tofu", [1, 1], [0.0, 2.0, 3.0]),
+        ("a cup of tea", [1, 0], [3.0]),
+        ("Is burrito my favorite food?", [0, 1], [1.0]),
+        ("I want club?mate", [1, 0], [2.0, 3.0]),
+    ],
+)
+def test_lookup_tables(sentence, expected, labeled_tokens, spacy_nlp):
+    from nlu.featurizers.sparse_featurizer.regex_featurizer import RegexFeaturizer
+
+    lookups = [
+        {
+            "name": "drinks",
+            "elements": ["mojito", "lemonade", "sweet berry wine", "tea", "club?mate"],
+        },
+        {"name": "plates", "elements": "data/test/lookup_tables/plates.txt"},
+    ]
+    ftr = RegexFeaturizer(lookup_tables=lookups)
+
+    # adds tokens to the message
+    tokenizer = SpacyTokenizer()
+    message = Message(sentence)
+    message.set("spacy_doc", spacy_nlp(sentence))
+    tokenizer.process(message)
+
+    result = ftr.features_for_patterns(message)
+    assert np.allclose(result, expected, atol=1e-10)
+
+    # the tokenizer should have added tokens
+    assert len(message.get("tokens", [])) > 0
+    # the number of regex matches on each token should match
+    for i, token in enumerate(message.get("tokens")):
+        token_matches = token.get("pattern").values()
+        num_matches = sum(token_matches)
+        assert num_matches == labeled_tokens.count(i)
